Make title smaller
Summary: Can't wait for i18n when this won't be a viable strategy any more.

Test Plan: In Chrome, open with ?sidebar=no, make the window 703px wide, then replace '0 of 750' with '750 of 750' and 'Download' with 'Downloaded' for maximum width. When making the pixel width smaller, it changes at around 407px so just do 405.

Reviewers: laura

Reviewed By: laura

Differential Revision: http://phabricator.khanacademy.org/D5032