Fix one of our video progress iOS bugs, this one triggered by play/pause/ended events when switching between videos.
Summary:
Since we recently switched to not reloading the entire video page when switching between videos, we haven't been hooking up appropriate events to the video player when switching. This means that if you load Video A, then switch to Video B, when you click on Start/Pause/Scrub on Video B you actually are sending /api/.../log events for Video A.

This was one of the sources of 'I just watched a video and no progress icons updated' ... but I don't think it's the only source.

...but no longer. I'd put an exclamation mark there if I knew how to escape it in my shell, but I can just never remember.

Test Plan:
On iPad
Log in
Load Video A
Switch to Video B
Click play, wait a few seconds, click pause
Verify that the progress icon for Video B immediately fills in

Reviewers: laura, alpert

Reviewed By: alpert

CC: mgp

Differential Revision: http://phabricator.khanacademy.org/D7169