Watching videos on iPad was inconsistently not giving full 750 point completion due to us 'losing' seconds of user progress here and there.
Summary: Make sure we don't spend time doing anything in our javascript code before we've updated dtSinceSave, otherwise all that time will be lost when we next send a /log request to the server.

Test Plan:
On iPad
Log in
Go to Partner Content => Getty => for kids (lots of short videos)
Watch lots of videos, and let them complete all the way to the end.
Verify that you receive 750 points for each and the progress icon is filled in (this doesn't happen consistently pre-patch).

Reviewers: laura, alpert

Differential Revision: http://phabricator.khanacademy.org/D7196