Remove the concept of offline state in the video view's JS state.
Summary:
We're doing this for two reasons: 1) it's currently tied up in the video view's OAuth logic, which we're about to abstract away to match the exercise view's (which paves the way for phantom users working in both video and exercise land), and 2) even without all this custom code for syncing video logging, our video logging already behaves sanely in the event of losing/regaining connection.

Specifically, if you lose a connection while trying to log progress and regain the connection a minute later, the /log API call that you send to the server will include all of that 'lost' time. We don't have to store these attempted saves and replay them the way this code attempts to do.

NOTE: there is a slight degredation here in that loading a video without a connection may fail to load subtitles, and when the connection is restored the subtitles will not re-load any more. Chalking this up to an improvement to be made later when we take a deeper look at making offline behavior nice later in the summer -- I don't think it's worth maintaining this old code and can be done in a much simpler way when needed.

Test Plan:
In simulator
Log in
Go to video you've never watched before
Click pause
Let the video buffer
Disable wifi on computer
Hook safari up to the simulator to watch requests go by
Click play, let the video play for 30 seconds
Verify that logging requests are sent, and fail, due to the lack of network connection
Reenable wifi
Verify that the next logging request sent contains a 'secondsWatched' value that represents all of the time you've spent watching the video, even while offline.
Verify that points earned are commensurate with the above.

Reviewers: alpert

Reviewed By: alpert

Subscribers: laura

Differential Revision: http://phabricator.khanacademy.org/D9446